[Refactor] Move AttributeService from Dashboard to Embeddable plugin
