+++ examples/embeddable_examples/public/book/edit_book_action.tsx
@@ -22,7 +22,11 @@ import { i18n } from '@kbn/i18n';
 import { BookSavedObjectAttributes, BOOK_SAVED_OBJECT } from '../../common';
 import { createAction } from '../../../../src/plugins/ui_actions/public';
 import { toMountPoint } from '../../../../src/plugins/kibana_react/public';
-import { ViewMode, SavedObjectEmbeddableInput } from '../../../../src/plugins/embeddable/public';
+import {
+  ViewMode,
+  SavedObjectEmbeddableInput,
+  EmbeddableStart,
+} from '../../../../src/plugins/embeddable/public';
 import {
   BookEmbeddable,
   BOOK_EMBEDDABLE,
@@ -30,13 +34,12 @@ import {
   BookByValueInput,
 } from './book_embeddable';
 import { CreateEditBookComponent } from './create_edit_book_component';
-import { DashboardStart } from '../../../../src/plugins/dashboard/public';
 import { OnSaveProps } from '../../../../src/plugins/saved_objects/public';
 import { SavedObjectsClientContract } from '../../../../src/core/target/types/public/saved_objects';
 
 interface StartServices {
   openModal: OverlayStart['openModal'];
-  getAttributeService: DashboardStart['getAttributeService'];
+  getAttributeService: EmbeddableStart['getAttributeService'];
   savedObjectsClient: SavedObjectsClientContract;
 }
