+++ examples/embeddable_examples/public/plugin.ts
@@ -62,7 +62,6 @@ import {
   ACTION_ADD_BOOK_TO_LIBRARY,
   createAddBookToLibraryAction,
 } from './book/add_book_to_library_action';
-import { DashboardStart } from '../../../src/plugins/dashboard/public';
 import {
   ACTION_UNLINK_BOOK_FROM_LIBRARY,
   createUnlinkBookFromLibraryAction,
@@ -75,7 +74,6 @@ export interface EmbeddableExamplesSetupDependencies {
 
 export interface EmbeddableExamplesStartDependencies {
   embeddable: EmbeddableStart;
-  dashboard: DashboardStart;
   savedObjectsClient: SavedObjectsClient;
 }
 
@@ -157,7 +155,7 @@ export class EmbeddableExamplesPlugin
     this.exampleEmbeddableFactories.getBookEmbeddableFactory = deps.embeddable.registerEmbeddableFactory(
       BOOK_EMBEDDABLE,
       new BookEmbeddableFactoryDefinition(async () => ({
-        getAttributeService: (await core.getStartServices())[1].dashboard.getAttributeService,
+        getAttributeService: (await core.getStartServices())[1].embeddable.getAttributeService,
         openModal: (await core.getStartServices())[0].overlays.openModal,
         savedObjectsClient: (await core.getStartServices())[0].savedObjects.client,
         overlays: (await core.getStartServices())[0].overlays,
@@ -165,7 +163,7 @@ export class EmbeddableExamplesPlugin
     );
 
     const editBookAction = createEditBookAction(async () => ({
-      getAttributeService: (await core.getStartServices())[1].dashboard.getAttributeService,
+      getAttributeService: (await core.getStartServices())[1].embeddable.getAttributeService,
       openModal: (await core.getStartServices())[0].overlays.openModal,
       savedObjectsClient: (await core.getStartServices())[0].savedObjects.client,
     }));
