+++ src/plugins/dashboard/public/plugin.tsx
@@ -39,8 +39,6 @@ import {
   CONTEXT_MENU_TRIGGER,
   EmbeddableSetup,
   EmbeddableStart,
-  SavedObjectEmbeddableInput,
-  EmbeddableInput,
   PANEL_NOTIFICATION_TRIGGER,
 } from '../../embeddable/public';
 import { DataPublicPluginSetup, DataPublicPluginStart, esFilters } from '../../data/public';
@@ -53,7 +51,6 @@ import {
   getSavedObjectFinder,
   SavedObjectLoader,
   SavedObjectsStart,
-  showSaveModal,
 } from '../../saved_objects/public';
 import {
   ExitFullScreenButton as ExitFullScreenButtonUi,
@@ -103,11 +100,6 @@ import { DashboardConstants } from './dashboard_constants';
 import { addEmbeddableToDashboardUrl } from './url_utils/url_helper';
 import { PlaceholderEmbeddableFactory } from './application/embeddable/placeholder';
 import { UrlGeneratorState } from '../../share/public';
-import { AttributeService } from '.';
-import {
-  AttributeServiceOptions,
-  ATTRIBUTE_SERVICE_KEY,
-} from './attribute_service/attribute_service';
 
 declare module '../../share/public' {
   export interface UrlGeneratorStateMapping {
@@ -156,16 +148,6 @@ export interface DashboardStart {
   dashboardUrlGenerator?: DashboardUrlGenerator;
   dashboardFeatureFlagConfig: DashboardFeatureFlagConfig;
   DashboardContainerByValueRenderer: ReturnType<typeof createDashboardContainerByValueRenderer>;
-  getAttributeService: <
-    A extends { title: string },
-    V extends EmbeddableInput & { [ATTRIBUTE_SERVICE_KEY]: A } = EmbeddableInput & {
-      [ATTRIBUTE_SERVICE_KEY]: A;
-    },
-    R extends SavedObjectEmbeddableInput = SavedObjectEmbeddableInput
-  >(
-    type: string,
-    options: AttributeServiceOptions<A>
-  ) => AttributeService<A, V, R>;
 }
 
 declare module '../../../plugins/ui_actions/public' {
@@ -433,7 +415,6 @@ export class DashboardPlugin
     const {
       uiActions,
       data: { indexPatterns, search },
-      embeddable,
     } = plugins;
 
     const SavedObjectFinder = getSavedObjectFinder(core.savedObjects, core.uiSettings);
@@ -483,15 +464,6 @@ export class DashboardPlugin
       DashboardContainerByValueRenderer: createDashboardContainerByValueRenderer({
         factory: dashboardContainerFactory,
       }),
-      getAttributeService: (type: string, options) =>
-        new AttributeService(
-          type,
-          showSaveModal,
-          core.i18n.Context,
-          core.notifications.toasts,
-          options,
-          embeddable.getEmbeddableFactory
-        ),
     };
   }
