+++ src/plugins/embeddable/public/lib/attribute_service/attribute_service.mock.tsx
@@ -17,11 +17,11 @@
  * under the License.
  */
 
-import { EmbeddableInput, SavedObjectEmbeddableInput } from '../embeddable_plugin';
-import { coreMock } from '../../../../core/public/mocks';
+import { EmbeddableInput, SavedObjectEmbeddableInput } from '../index';
+import { coreMock } from '../../../../../core/public/mocks';
 import { AttributeServiceOptions } from './attribute_service';
-import { CoreStart } from '../../../../core/public';
-import { AttributeService, ATTRIBUTE_SERVICE_KEY } from '..';
+import { CoreStart } from 'src/core/public';
+import { AttributeService, ATTRIBUTE_SERVICE_KEY } from './index';
 
 export const mockAttributeService = <
   A extends { title: string },
