+++ src/plugins/visualizations/kibana.json
@@ -3,7 +3,7 @@
   "version": "kibana",
   "server": true,
   "ui": true,
-  "requiredPlugins": ["data", "expressions", "uiActions", "embeddable", "inspector", "dashboard"],
+  "requiredPlugins": ["data", "expressions", "uiActions", "embeddable", "inspector" ],
   "optionalPlugins": ["usageCollection"],
   "requiredBundles": ["kibanaUtils", "discover", "savedObjects"]
 }
