+++ src/plugins/visualizations/public/embeddable/visualize_embeddable.ts
@@ -38,6 +38,7 @@ import {
   Adapters,
   SavedObjectEmbeddableInput,
   ReferenceOrValueEmbeddable,
+  AttributeService,
 } from '../../../../plugins/embeddable/public';
 import {
   IExpressionLoaderParams,
@@ -51,7 +52,6 @@ import { VIS_EVENT_TO_TRIGGER } from './events';
 import { VisualizeEmbeddableFactoryDeps } from './visualize_embeddable_factory';
 import { TriggerId } from '../../../ui_actions/public';
 import { SavedObjectAttributes } from '../../../../core/types';
-import { AttributeService } from '../../../dashboard/public';
 import { SavedVisualizationsLoader } from '../saved_visualizations';
 import { VisSavedObject } from '../types';
